Possible fix for non Windows OS
